requirements.txt
python-dotenv=0.10.1
requests=2.18.4